# services/gen3_embeddings/docs/autogenerated_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gen3 Embeddings Service",
    "version": "0.1.2"
  },
  "paths": {
...
    "/vectorstore/collections": {
      "get": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Read all collections",
        "description": "Returns every vector collection you have access to, one page at a time. Set `counts=true` to also get the number of embeddings available in each collection.",
        "operationId": "list_collections_vectorstore_collections_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          },
          {
            "name": "counts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Counts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCollectionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Create collection",
        "description": "Creates a new vector collection. The collection name is normalized before it is stored, and the dimensions you supply here are enforced on every embedding added to the collection.",
        "operationId": "create_collection_vectorstore_collections_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModel"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vectorstore/collections/{collection_name}": {
      "get": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Read collection info",
        "description": "Returns the metadata for a single collection, including its dimensions and vector type. Set `counts=true` to also get the number of embeddings available in it.",
        "operationId": "get_collection_vectorstore_collections__collection_name__get",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "counts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Counts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModel"
                }
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Update collection info",
        "description": "Updates the mutable metadata on a collection, such as its description. The collection's name, dimensions, and vector type cannot be changed after creation.",
        "operationId": "update_collection_vectorstore_collections__collection_name__patch",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vectorstore Collections"
        ],
        "summary": "Delete collection",
        "description": "Permanently deletes a collection and every embedding stored in it. This cannot be undone.",
        "operationId": "delete_collection_vectorstore_collections__collection_name__delete",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation, no content returned"
          },
          "401": {
            "description": "User unauthenticated"
          },
          "403": {
            "description": "User does not have access"
          },
          "400": {
            "description": "Invalid request payload or parameters"
          },
          "404": {
            "description": "Collection not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CollectionModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dimensions": {
            "type": "integer",
            "title": "Dimensions"
          },
          "vector_type": {
            "$ref": "#/components/schemas/VectorType"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "self": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self"
          },
          "available_embeddings_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Embeddings Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "collection_name",
          "dimensions",
          "vector_type"
        ],
        "title": "CollectionModel",
        "description": "API schema representing a collection."
      },
      "CreateCollectionBody": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dimensions": {
            "type": "integer",
            "title": "Dimensions"
          },
          "vector_type": {
            "$ref": "#/components/schemas/VectorType",
            "default": "vector"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "dimensions"
        ],
        "title": "CreateCollectionBody",
        "description": "Request body for creating a new Collection."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaginatedCollectionsResponse": {
        "properties": {
          "collections": {
            "items": {
              "$ref": "#/components/schemas/CollectionModel"
            },
            "type": "array",
            "title": "Collections"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "next_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page"
          },
          "prev_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Page"
          }
        },
        "type": "object",
        "required": [
          "collections",
          "page",
          "page_size"
        ],
        "title": "PaginatedCollectionsResponse",
        "description": "A page of collections the caller is authorized to see."
      },
...
      "UpdateCollectionBody": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "UpdateCollectionBody",
        "description": "Request body for updating mutable properties of a Collection."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VectorType": {
        "type": "string",
        "enum": [
          "vector",
          "halfvec"
        ],
        "title": "VectorType",
        "description": "pgvector column type backing a collection, which fixes its storage precision."
      }
    }
  }
}